requirements.txt
requests==2.28.1
python-binance==1.0.16
python-dotenv==0.20.0